module5-task1: * move game redux flow
