+++ src/store/actions/game/action-type.ts
